New version. Tiny bug fix, Ubuntu based distros only.
The 3.0.13 system base feature had a small bug in the logic that was supposed to
get the version id from codename, the bug made it never work. This is only relevant
for Ubuntu based distros, so if you are on some other base like Debian or Arch, you
can ignore this one, 3.0.13 will work fine.

No other changes, this was mainly for Mint, and other Ubuntu derived distros in
the future.