New version, new man. Big bug fix, new features.
Bugs:
1. Finally tracked down and solved the Xorg drivers bug which was caused by
Xorg checking its list of defaults 2 times, not 1, which resulted in failed
status on second try since it was already loaded. Secondary bug was found that
resulted in failing to show the failed, and only showing unloaded, which was also
wrong. This finally fixes issue #134 item 5. Thanks Mint users for the help on
that one.
2. Small bug in Openbox version detection, typo.
3. fixed a small glitch in the dm: detection that on systems where /var/run
exists but is not linked to /run, the dm would fail to get detected.

Fixes:
1. Xfce when defaulting to no version found goes to 4, this is a bad idea, it's
better to not show any version, since xfcie could one day be 5.
2. Fixed Blackbox fallback detection, there  were cases where Blackpox not found
in xprop -root, now it falls back to ps aux detection.
3. For wm: tested all known variants, added support for things like Mutter (Marco)
syntax. Note that bunsenlab uses XDG_CURRENT_DESKTOP=XFCE to work around some
glitches, but it's actually Openbox. If run as root, it will show openbox correctly,
otherwise -Sxxx will show wm: openbox, but that's due to bunsenlabs choices there.
4. Rewrote a lot of DistroData to handle more dynamic testing of values, it's sad
that at almost 2020 we are still stumbling around trying to find a consistent way
to identify distros, and derived distros.
5. Added more debugger data collectors in the logging, some data was not
being tracked well during log process which made debugging harder.

Enhancements:
1. New feature, -Gxx now shows for Xorg drivers alternate: which are drivers that
Xorg auto checks but which are not installed. Those were ignored in the past. This
can be useful to see for example that there are other driver install options
available. Thanks gm10 for that suggestion.
2. Tested and added the following explicit handlers for Distros: and base: in
some cases:
grml, peppermint, kali, siduction, aptosid, arco, manjaro, chakra, antergos,
bunsenlabs, and a few others. These are a pain to add and test, basically I have
to boot a live cd of each one, then test the files and ID methods, but the ID
methods must also be as dynamic as possible because you never know when a distro
is going to change how they use os-release vs issue vs lsb-release vs <name>-release.
I would have tested a few more but their livecds failed to properly run on vbox.
3. Added a few more disk vendor IDs.
4. Added some more programs to debugger data collector for future feature vdpau, but
that needs more data because we don't really know the variants for example for
dual card systems.
5. Man page: changed extra options to use only one option name per list of options
for that feature, each separe item is started as a new paragraph with -
This makes it a bit more consistent and maybe slightly easier to read the man.
Added -Gxx item, updated -Sx item.