New version, new man. Changes, bug fixes, enhancements! Don't delay!
Bugs:
1. A real bug, the detection for true path of /dev/root had a mistake in it and
would only have worked in half the cases. This was an easy fix, but a significant
but since it also would lead to the actual root / partition showing in Unmounted.
2. Related to the item Fixes-2, if two USB networking devices  were attached,
the second one's bus and chip ID would go on the wrong line of data if -n or -i
option were used. Since that would be the line belonging to the one above,
that obviously was weird and wrong.
3. NEW: latest kernel can show hwmon data in sensors, for example from wifi chip.
This broke CPU temp detection and showed way too high cpu temp, so this fix is
fairly important since new kernels may have this new sensors hwmon syntax.
4. Sensors: IPMI alternate syntax found, also case with no data in expected columns,
just N/A, so now the ipmi sensor logic skips all lines with non numeric values in
the values column. This is what it should have done all along, it was trusting
that values would always exist for the field names it looks for.

Fixes:
1. ARM networking fix. ARM devices like rasberry pi that use usb bus for networking
showed the no data message even though usb networking was right below it. This is
corrected, and now that only shows if both main and usb networking failed for ARM.
2. Big repo fix: while testing distro and Trinity live cds, I discovered that apt is
sometimes used with rpms, which made PCLinuxOS and ALT-Linux Repos item show the
apt files but no data since the pattern was looking for start with deb. Added rpm
to pattern, so all distros that use apt running rpms should now 'just work'.
3. Fixed more distro id things, PCLinuxOS should now show its full distro string.
4. Debugger: Filtered out more blocks of /proc, that data is bloated and messy, found
another case where it collected a vast amount of junk system data from zfs in that
case, just blocked the entire range. I had no idea /proc had so much junk data in it!
5. As noted above, IPMI, yet another alternate syntax for field names. My hope that
IPMI software and sensors will be more logical and consistent than lm-sensors output
is proving to be merely wishful thinking, I think now out of 3 datasets I've gotten,
I've seen 3 variants for syntax, not to mention the ipmi-tool vs ipmi-sensors
differences. So IPMI will be like all sensors stuff, a work in progress, to
be updated with every newly discovered alternate syntax and data set.

Enhancements:
1. Disk vendors, added some, improved pattern detections for others. This feature
is getting better all the time. Thanks linuxlite hw db, easy to scan for missing
vendors in their inxi data.
2. Added more wm, budgie-wm, mwm, variants of kwin and Trinity's Twin, several others,
more refactoring of core wm/desktop code.
3. Added gpu ram and reworked memory logic for rasberry pi, which is the only SBC
I am aware of that uses that tool. Now reports the actual total, and also gpu: for
ram data, so you can tell that the gpu is using part of the total. Again, this comes
from issue #153. Also added that info to man page for -I part.
4. Added more ARM and PCI cleaners for neater and more concise ARM/PCI output.
5. Added Trinity support to Desktop section, this had at least two different detection
methods, but since the first just shows KDE original data, only the second one proved
to be Trinity specific. Happily, the full data is available, toolkit, desktop version,
and wm (Twin).
6. New -G,-A,-R -xxx feature: vendor:. Note that vendor data is very bloated and
messy so it's trimmed down substantially, using a series of filters and rules, and
thus it can contain the following: the actual vendor, like Dell, nothing, the
motherboard vendor/product for board based PCI items, or a complete vendor/product
string if it's unique. I couldn't think of a clean field name that meant:
vendor OR vendor + basic product info OR motherboard + board version OR full
product name, including vendor, so in the end, I just used vendor: but it's not
quite the right term, but nothing else seemed to work better. Testers responded
very enthusiastically about this feature so I guess the vendor: name is ok.

Changes:
1. Biggest change: Drives: HDD: total: the HDD: is now changed to: Local Storage:
This was part of issue #153 and is a good suggestion because HDD generally was used
to refer to hard disks, spinning, but with nvme, m.2, ssd, etc, that term is a bit
dated. 'Local' is because inxi does not include detected remote storage in the totals.
2. The recent --wm option which forced ps as data source for window manager detection
has been reversed, now --wm forces wmctrl and ps aux is preferred. Still falls back
to wm ctrl in case the ps test is null, this is better because I have to add the wm
data manually for each one, whereas wmctrl has an unknown set and probably variable
set of wm. Note that I reversed this because I saw several cases where wmctrl was
wrong, and reported a generic source wm instead of the real one. Since most uses are
not going to even be aware of the wm: feature as enhanced with --wm switch, this
should have no impact on users in general. Since the detected wm name needs to be
know to get assigned to wm: and wm version data, I think it will work better to
have the known variants match with the wm data values, then just fallback to
unknown ones that can get fille in over time as we find wm that people actually
use and that you can get version info on and detect.

Removed:
1. Got rid of tests for GTK compiled with version for many desktops, that test
was always wrong because it did not have any necessary relation to the actual
gtk version the desktop was built out of, and it also almost always returned no
data. Since this is an expensive and slow test, and is always going to be wrong
or empty anyway, I've removed it. My tests showed it taking about 300ms or so
to generate no data, heh.
That's the tk: feature in -S. Note I also found that gnome-shell takes
an absurdly long time to give --version info, the slowest of all such things, 300ms
again, just to show version? Someone should fix that, there's no possible reason
why it should take 300 milliseconds to give a simple version string. Note that
this returns tk: to only returning real data, which in this case means only xfce,
kde, and trinity, which are the only desktops that actually report their toolkit
data. I'll probably remove that code in the future unless I can think of some real
use for gtk version elsewhere, but it's just junk data which doesn't even work.

In the future, I will not try to emulate or guess at desktop toolkits, either they
show the data in a direct form like XFCE or Trinity or KDE do, or I won't waste
resources and execution time making bad guesses using inefficient code and logic.
QT desktops like LXQt I'm leaving in because I believe those will tend to track
more closely the QT version on the system, and the tests for QT version aren't
huge ugly hacks the way they are for GTK, so they aren't as slow or intrusive, but
those may also get removed since they almost never work either. But they are also
slowing down the -Sx process so maybe they should be removed as well, I'll think
about it. Since they only are used on LXQt and razer-qt, it probably isn't a big
deal overall.