New version, new man. Fixes, glitches, and stitches!
Fixed some subtle and not subtle issues that I've noticed recently.

Bugs:
1. The color scheme selector failed to remove the global value when a non global
setting was used. This led to global values never getting removed, even though
the text output said it would be, which is confusing, obviously, and always
overriding the color selected. Thanks CentOS for helping find that one.

Fixes:
1. Fixed possible corrupted user inxi.conf values. Now skips null values, and
fully validates as integer integer values.
2. Fixed fvwm-crystal detections, integrated it into new refactored desktop logic.
3. For systems without glxinfo or running inxi out of gui/desktop, Xorg was in many
cases failing to show version, which made it not show anything for server: except
N/A. This is caused by a relatively recent change in behaviors in xorg, where you
have to run it directly from it's true path, which is something like /usr/lib/xorg
or /usr/lib/server-xorg at which point the error:
/usr/lib/xorg-server/Xorg.wrap: Only console users are allowed to run the X server
Figuring this out was tricky, and who the heck knows why Xorg -version would even
return such a silly error in the first place, but there you have it. Next time
you wonder why inxi is so long, this is why, endless churn in basic and complex
things! The fix is injecting the optional xorg paths into @paths right before,
and removing them right after, which avoids adding clutter to the @paths.
4. A ZFS fix, I'd noticed this one a while back, but after looking at the zfs
Ubuntu tutorial page, I realized that this is the norm now, which is building zfs
with /dev/sda (no partitions). This lead to failing to detect the zfs components,
and reporting a bunch of partitions as unmounted which were part of that /dev/sdb
type component array. By allowing /dev/sd[a-z] I fixed both errors at the same time,
but I don't know if this syntax extends to say, nvme zfs as well. Note that when
you build zfs arrays with say, /dev/sdb /dev/sdc you'll see two partitions per
disk, /dev/sdx1 which is the main data, and /dev/sdx2, which is a tiny 8mB partition,
no idea what it's for.
5. Fixed missing konversation and hexchat version numbers in -I, finally found
what was going on there. Note that hexchat --version used to pop up a gui, but
I guess he finally fixed that, I am hoping.
6. Fixed some gentoo repo detections, but also found more variants. Not sure what
exactly is going on with repos there, will wait for gentoo user issue reports to
really lock those down.
7. BSD fixes, turns out FreeBSD uses that same map ... syntax in df -kT as OSX...
Also made sure to load sysctl data for -S row, I'd forgotten about the compiler
test there which needs that data.
8. Fixed herbstluftwm version detection, turns out it's another one of those that
passes the entire path to the version program, so it shows: /sbin/herbsuftwm 0.22.0
which broke the regex, easy fix.
9. Completed refactoring of DesktopData, now it's all data array driven for most
wm, desktops, etc, which makes adding/removing one very easy. All core data is now
in program_values to allow for automated detections.

Enhancements:
1. With fix 1, added check_int and check_number utilities, these validate that inxi
internal numeric or integer values actually are what they are supposed to be. This
uses a neat Perl trick that makse the checks super fast and super accurate. Moved
all internal int/numeric test regex to use these.
2. Added file based version number detection, that was done for Deepin, which uses
/etc/deepin-version for its version number, but it can be used for anything.
3. Added Deepin and deepin window managers, Lumina, added bspwm wm, fixed muffin
detections. Note that lumina has a weird behavior where when run outside of pinxi,
it outputs to stdout, but inside of pinxi, to stderr, who the heck knows how that
happens!
4. Added zorin to supported base: distros.
5. Even more disk vendors added! The list of no-name off brand chinese ssd vendors
appears to be endless! Added some more specific ids to capture unique strings
that can be linked to a vendor.
6. Added /usr/home to default -P paths, that's used instead of /home in the real
world, so why not show it?
7. Because qt detection is possible, I've extended qt toolkit detection, but it's
also not super accurate, but it's far better than gtk tk was, so I'm leaving
that in. I also extended it to more wm/desktops since more are using qt now.
Note: budgie 11 is going to be qt, but there's no way to distinguish between 11 and
gtk 10 without doing a bunch of hacks so I'm leaving that alone.
8. Found a possible distro id source, added /etc/calamares detections to debugger,
I'll see if that shows some consistent patterns before I implement a last fallback
test for distro IDs. It may work.

Removed:
1. Giving up on fake/slow/inaccurate GTK toolkit detections, removed the entire
codeblock and stored in docs/inxi-fragments.txt, but I'm not going to do package
manager type version tests anymore, if we can't get the data directly from a program
or file, it's not going to happen, plus the gtk installed on the system means nothing
in relation to the gtk version used to build the desktop.