New version, man page. Bug fixes, enhancements.
Bugs:
1. A long standing bug was finally identified and fixed. -n/-i would fail to match
a Device to the right IF in cases where they had the same chip / vendor IDs. Added
busID for non Soc type devices to fix that. I hope. This fix has been tested on a
machine that had this bug, and it is now corrected. Thanks skynet for the dataset.
2. deepin-wm was failing to get listed correctly with new fixes, this is corrected.

Fixes:
1. mate version was depending on two tools, mate-about and mate-session, which
somewhat randomly vary in which has the actual highest version number. Fix was to
run both in MATE for version, and run those through a new version compare tool.
Thanks mint/gm10 for reporting that bug.
2. -Gxx compositors: added some missing ones that were being checked for in-
correctly.
3. For distro id, fixed a glitch in the parser for files, now correctly removes
empty () with or without spaces in it.
4. Got rid of ' SOC?' part of no data for ram or slots, that also triggers in non
SOC cases, so best to not guess if I can't get it right.

Enhancements:
1. More disk vendor ID matches, also, somehow missed QEMU as vendor, thanks to
linux hardware database (linuxlite) for great samples of vendor/product strings.
2. Added a bunch of compositors, found a new source that listed a lot inxi did not
have already.
3. Added version v: for some compositors in -Gxxx.
4. New program_data() tool provides an easier to use simple program version/print
name generator, including extra level tests, to get rid of some code that repeats.
5. Found some useful QEMU virtual machines for ARM, MIPS, PPC, and SPARC, so
made initial debugging for each type, so basic working error free support is well
on its way for all 4 architectures, which was unexpected. More fine tunings to
all of them to avoid bugs, and to catch more devices, as well.
Note that QEMU images are hard to make, and they were not complete in terms of
what you would see on physical hardware, so I don't know what features will work
or not work, there may be further variants in audio/network/graphics IDs that
remain unhandled, new datasets always welcome for such platforms!
6. Found yet another desktop! Added Manokwari support, which is at this point
a reworking of gnome, but it was identifiable, minus a version number.
7. Added deepin and blankon to system base supported list, these hide their debian
roots, so I had to use the manual method to provide system base.