New version, new man.
Bugs:
1. If you consider failure to identify a mounted yet hidden partition a bug, then
that bug is fixed, but I consider that as more of a fix than a bug.

Fixes:
1. Added more device pattern ID for odroid C1 and C2, these are now pretty well
supported.
2. inxi failed to handle a certain type of hidden partition, so far only seen
with udiskctl mounted TimeShift partitions, but this may be a more general udisk
issue, but so far not enough information. The fix is to use the lsblk data to
build up missing partitions, so this fix is for non legacy Linux systems only.
The fix works pretty well, but it's hard to know until we get a lot more real
world data, but given so far I've received only one issue report on it, I
suspect this is not a common situation, but you never know, it would never
have shown up in datasets unless I had looked specifically for it, so it may
be more common than I think.
3. Cleaned up and simplified new --admin -p and -d logic.

Enhancements:
1. For debugging, renamed all user debugger switches to have prefix --debug.
These options are to help debug debugger failures, and so far have been tested
and solved the failures, so I'm adding them all to the main man and help menu,
thus raising them to the level of supported tools. These were enormously helpful
in solving proc or sys debugger hangs.
 * --debug-proc
 * --debug-proc-print
 * --debug-no-sys
 * --debug-sys
 * --debug-sys-print
2. Added findmnt output to debugger, that may be useful in the future. Also added
df -kTPa to also catch hidden partitions in debugger.
3. Added in another user level debugger, triggered with --debug-test-1 flag. This
will do whatever operation is needed at the time for that user. Some issues can
only be resolved by the user on their machine.
4. More disk vendors and matches!!! Thanks linuxlite/linux hardware database!