New version, new man. Fixes, stitches, and returns!!
Bugs:
1. As a fix (3), failure to handle spaces in mount source names. More of a fix than
a bug, since it was an old issue #63.
2. OSX errors, BSD errors, but not really inxi errors or bugs, more weird data
tripping null data or unreadable file errors, but I'll call those bugs since they
look like bugs to end users. See Fixes for more.
3. See Fix 4, this is sort of a bug, inxi failed to return expected values on
success/failure.

Fixes:
1. One of the documented config items, COLS_MAX_NO_DISPLAY had not been implemented
internally. This is now corrected.
2. Apple OSX was returning errors, those were fixed.
3. Finally handled ancient issue #63, support now there for spaces in remote source
name. This means that both spaces in source block name, and mount point name, are
in theory both handled now. This was also to fix an osx issue #164 despite the
fact that technically I do not support osx beyond fixing errors, but since in this
case the issue was a long standing one, I fixed it for everything.
4. Big fix, I'd completely left undone proper unix type error number returns in
inxi, oops. Thanks Unit193 for noticing that and prompting me to fix it. Now inxi
returns integer success/error numbers as expected.
5. OSX xml based version info broke, of course, naturally it would, so I added in
an osx fallback where if no distro version detected, use fallback unix method, which
is what all the other unices use.
6. Along with space in source name, fixed mapped handling a bit more too for
partitions.
6. Added cifs remote file system to disk size used blacklist, and iso9660. Not sure
how I'd missed those for so long.
7. OpenBSD vmstat in 6.3 changed the column order for avm/fre, and changed to a,
sigh, human readable default format, in M, so to handle this for all bsds, I had
to make a dynamic column detection for avm and fre, and use those after, and also
i had to add in a M detection, if found, *1024 and strip out M, sigh.
8. OpenBSD, another alternate ordering/syntax issue, the dmesg.boot data for disks
does not always use the same order in comma separated list, saw user case where
the first item after : was the MB size, not the second. Made detection dynamic.
9. Due to Android case, found types where no cpu speed data was found, no max speed
at least, which tripped an error due to null data for ARM, this is now handled,
now cpu speed min/max read permissions in /sys are checked first before trying to
read, and default failures are better handled.
10. On man page, added in clarification of the moving of Memory: item from Info:
line to ram Memory: line, explaining when it appears where. I do not removing the
item from -I, I may revert that change, I find it non-intuitive to move that
around.

Enhancements:
1. Added display manager Ly, plus Ly version number. Thanks NamedKitten, this
closes issues #166 #165 #162
2. Improved documentation a bit to avoid ambiguity re how to get colors in output.
That handles issue #161, thanks fugo for the nudge to improve the documentation.
3. First inxi on Android tests, using termux, which has a debian based apt
type installer, got inxi running on at least two devices, including pixel2, but
discovered that apparently as of android 5, /sys is now locked up in terms of
wildcard reads, but further analysis is required, but as of now, inxi works in
termux, but fails to get any Device data for A, G, or N. Thus it also fails to match
IF to Device, so none of the IP data shows up. The latter will probably be fixed
since Android has ip and ifconfig already, or termux does, but so far I found no
way to get device data for ARM in Android 5.x and greater (checked on
android 7 and 9 in real phones).
4. More disk vendors!! thanks linuxlite / linux hardware database for offering an
apparently never ending list of obscure and not so obscure disk vendors and
products.
5. While I was unable to get confirmation or documentation on file names for
tce repo files, I guessed that localmirrors would be used, but this may be
any random text file in /opt at all, no extensions, I'd have to test to confirm
or deny possible values.
6. To handle more complex debugger failures, added --debug-no-proc,
--debug-no-exit, to skip or enable completion where proc or sys debugger is hanging.

Changes:
1. Changed vendor in A, G, and N to -x, not -xxx, this data seems much more useful
and reliable than I'd first expected when I made the feature, the -xxx was more
an indication of my lack of trust in the method and source, but so far it seems
pretty good, so I bumped it up to an -x option. Note that also, it's quite useful
to know the vendor of, say, your network or graphics card, not just the actual
device internal data, which is all inxi has ever shown previously.
2. Small change, if no partition type data is found, dev, remote, mapped, default
now says 'source:' instead of 'dev:' which makes more sense. Note that df calls
that column 'source', so I decided to go with their language for the default not
found case. Also changed mapped to say mapped. This was part of a bit of a
refactor of the partition type logic, enhanced by adding mapped to existing types,
and moved the entire type detection block into the main data generator, and out
of the data line constructor.

Optimizations:
1. Tested, and dumped, List::Util first() as a possible way to speed up grep
searches of arrays, where the goal is just to see if something is in an array. My
expectation was that first(), returning the first found instance of the search term,
would of course be faster since it will always exit the search loop was met with
the sad fact that first() is about 2 to 4 times SLOWER than grep() native builtin.
I tested this fairly carefully, and used NYTProf optimizer tool and the results were
totally consistent, first() was always much slower than grep(), no matter what size
the array is. I assume this means the core C programming that makes grep is simply
much better than the File::Util module programming that makes first(). Removed
first() and now know that nothing will be faster than grep so no need to look there
for speed improvements.
The moral of the story: just because something should in theory be faster, does
sadly not mean it will be faster, for there are bigger things at work, skill of
the programmers who made the logic, how perl handles external vs internal tools,
and so on. As an aside, this forms a fairly consistent pattern where I've found
Perl by itself to be faster than modules in many cases, that is, it's faster to
write the code out than to use a module in many cases that I have tested, so I
will always test such ideas and dump every one that is in fact slower than native
Perl builtins.