New version, new man. Fixes, enhancements.
Bugs:
1. ARM fix, odroid > 1 cpu speeds not showing correctly.
2. Ansible start fixes.
3. Fringe Battery failures, see Pinebook.

Fixes:
1. Removed null data message 'old system' since that's not always the case.
2. Added support for > 1 CPU speeds in systems with > 1 CPU.
3. Added is_numeric test for sudo version tests, that was tripping errors in
rare cases.
4. Fine tuned terminal size setting to check that is int to correct the
Ansible problem.
5. ARM Pinebook fixes, battery, cpu. This also fixes corner cases where
the battery charge state is missing but it is a systme battery.

Enhancments:
1. Added more disk ID matches/vendors. Thanks LinuxLite Hardware database!!
2. UKUI, ukwm, ukui-panel added to desktop data.
3. Added PopOS to system base.
4. Ansible/Chef user noted that inxi believes that it is running in IRC when
started by Ansible / Chef (not sure about Chef but assuming it's the same).
Added flag --tty flag to force inxi to believe it's running in shell no matter
what starts it. Note that this fix is not confirmed because the person didn't
confirm the fix. Annoying.
5. Added Ubuntu disco to ubuntu_id.