New version, new man page.
Bugs:
1. Both a fix and a bug, in that inxi had an out of date list of Xorg drivers.
This led to all the newer Intel devices failing to show their drivers in the
Xorg driver lines, like i915, i965, and so on. Updated to full current list of
Xorg drivers. This is not technically a bug since it's simply things that came
into existence after that logic was last updated. But it looks like a bug.

Fixes:
1. Issues #170 and #168 showed a problem with inxi believing it was running in IRC
when Ansible or MOTD started inxi. This is because they are not tty so trip the
non tty flag, which assumes it's in IRC in that case. The fix was to add a
whitelist of known clients based on the parent name inxi discovers while running
inside that parent. MOTD confirmed fixed, Ansible not confirmed. Why do people file
issue reports then not follow them? Who knows.
Note that this issue is easy to trip by simply doing this: echo 'fred' | inxi
which disables the tty test as well. To handle that scenario, that is, when inxi is
not first in the pipe, I added many known terminal client names to the whitelists.
This works in my tests, though the possible terminals, or programs with embedded
terminals, is quite large, but inxi handles most of them automatically. When it
doesn't, file an issue and I'll add your client ID to the whitelist, and use --tty
in the meantime.
2. Issue #171 by Vascom finally pinned down the wide character issue which manifests
in some character sets, like greek or russian utf8. The fix was more of a work-around
than a true fix, but inxi now simply checks the weather local time output for wide
characters, and if detected, switches the local date/time format to iso standard,
which is does not contain non ascii characters as far as I can tell. This seemed to
fix the issue.
3. Added iso9660 from excluded file systems for partitions, not sure how inxi
missed that one for so long.
4. See bug 1, expanded and made current supported intel drivers, and a few other
drivers, so now inxi has all the supported xorg drivers again. Updated docs as well
to indicate where to get that data.

Enhancements:
1. As usual, more disk vendor/product ID matches, thanks to linuxlite hardware
database, which never stops providing new or previously unseen disk ids. Latest
favorite? Swissarmy knife maker victorinox Swissflash usb device.
2. Added Elive system base ID.
3. Added Nutyx CARDS repo type.