New version, new man page. Big update! Get it in before your freeze!!
Bugs:
1. Maybe the vendor/product regex, which when + was used, would put out
errors.
2. Maybe Fix 4, since that could lead to incorrect behavior when sudo
is involved depending on sudo configuration.
3. BIG: current inxi weather will probably fail if not updated to this or
newer versions!! Not an inxi bug per se, but your users will see it as one.

Fixes:
1. Fixed Patriot disk ID.
2. Fixes for PPC board handling.
3. Regex cleaner fixes, this could lead to error in special cases of product
vendor names.
4. crazy from frugalware pointed out that $b_root detection was flawed, and
relied on a bad assumption, particularly for sudo. As usual, he's right, that
is now corrected, and uses $< Perl native to determine UID.

Enhancements:
1. Added septor to Debian system base.
2. Removed quiet filters for downloaders when using --dbg 1, now you see the
entire download action for curl/wget downloads. This went along with
issue # 174
3. New feature: --wan-ip-url. This closed issue #174. Also has user config
option: WAN_IP_URL as well to make changes permanent.
4. Added --dbg 1 to man and help. The other --dbg options are random and can
change, but --dbg 1 is always for downloading, so might as well tell people
about it.
5. To anticipate the loss of a major weather API, inxi is redone to use smxi.org
based robust API. This also allows for a new switch, --weather-source (or --ws
for shorter version), options 0-9, which will trigger different APIs on smxi.org.
Added WEATHER_SOURCE configuration option as well. Note that 4-9 are not
currently active. Also added in better error handling for weather.
The main benefit here is that inxi is now largely agnostic to the weather APIs
used, and those can be changed with no impact to inxi users who are running
frozen pool inxi's, or who have not updated their inxi versions.

NOTE: all inxi versions older than 3.0.31 will probably fail for weather
quite soon. So update your inxi version in your repos!!
6. More disk vendors IDs and matches. Thanks linuxlite hardware database.
7. Going along with weather changes, added, if present, cloud cover, rain, and
snow reports. Those are for previously observed hour.
8. Small change to Intel CPU architecture, taking a guess on stepping for
skylake/Cascade lake ID. Guessing if stepping is > 4, it's cascade lake. But
could not find this documented, so it's a guess. At worst, it  means that Cascade
lake, which must be a later steppingi than 4, will not be ID'ed as skylake.
9. Documentation updates for data sources.

Changes:
1. inxi now uses a new system to get weather data. There is no longer a risk
of weather failing if the API used locally in inxi fails or goes away. This
change should be largely invisible to casual users.
2. In weather, moved dewpoint to be after humidity, which makes a little more
sense.