New version, new man, new feature!! Bug fixes!
Bugs:
1. issue #182 - in freebsd, there was an oversight in the pciconf parser, it
was using unfiltered strings as regex pattern, and of course, a string flipped
an error. Fix was to add the regex cleaner to the string before it's used in test.

2. NOTE: issue #182 had a second bug, but the issue poster didn't follow up with
data or output so it couldn't be fixed. This was related to a syntax change in
usbdevs -v output in FreeBSD. Such changes are too common, but it might also
simply be a variant I have not seen or handled, but so far no data, so can't fix.
Don't blame me if you get this bug, but do post requested debugger data if you
want it fixed!

Fixes:
1. Updated man for weather, explained more clearly how to use country codes for
weather output. More clarifying in general about weather location, and weather
restrictions.

Enhancements:
1. Added avx/avx2 to default flag list in -C short form. Thanks damentz from
liquorix for clarifying why that was a good idea. Note the initial issue came up
in a Debian issue report, not here. People!! please post issues here, and don't bug
maintainers with feature requests! Maintainers aren't in a position to add a feature,
so you should go straight to the source.

1.a. Created in inxi-perl/docs new doc file: cpu-flags.txt, which explains all
the flags, and also covers the short form flags and explains why they are used.

2. To resolve another issue, I made a new documentation file:
inxi-perl/docs/inxi-custom-recommends.txt
This is instructions for maintainers of distros who do not use rpm/apt/pacman but
still want the --recommends feature to output their package pool package names for
missing packages. I decided to not allow more than the default 3 package managers
because no matter what people say, if I allow in more, the maintainer will vanish
or lose interest, and I'll be stuck having to maintain their package lists forever.

Also, it's silly to even include that package list for any distro that does not
use rpm/apt/pacman, since the list is just wasted lines. Instructions in doc file
show what to change, and how, and has an example to make it clear. Odds of this
actually being used? Not high, lol, but that's fine, if people want it done, they
can do it, if not, nothing bad happens, it just won't show any suggested install
package, no big deal.

3. Using the new disk vendor method, added even more disk vendors. Thanks
linux litet hardware database!!

4. EXCITING!! A new --admin/-a option, suggested by a user on techpatterns.com/forums/
Now -S or -b or -F with -a option for GNU/Linux shows the kernel boot parameters,
from /proc/cmdline. Didn't find anything comparable for BSDs, if you can tell me
where to look, I'll add it for those too, but wasn't anywhere I looked. Do the
BSDs even use that method? Don't know, but the logic is there, waiting to be used
if someone shows me how to get it cleanly. The 'parameters:' item shows in the main
'System:' -S output, and will just show the entire kernel parameters used to boot.

This could be very helpful to distros who often have to determine if for example
graphics blacklists are correctly applied for non free drivers, like nomodeset etc,
or if the opposite is present.

For forum/distro support, they just have to ask for: inxi -ba and they will see t
the relevant graphics info, for instance, or -SGaxxx, or -Faxxx, whatever is used
to trigger in this case the graphics and system lines.

5. Updated man/help for 4 as well, now explains what they will see with --admin/
-a options and -S. Good user suggestion, I wish all new features were this easy,
heh.