New version, many small fixes. And a hall of shame, LOL.
Bugs:
1. Issue #188 exposed a situation in glxinfo where the required opengl fields are
present but contain null data. This happens when a system does not have the required
opengl drivers, which was the case here. inxi failed to handle that. Thanks
LinuxMonger for posting the required data to figure this corner case out.

2. Fixed a long time bug in Disk vendor ID, there was an eq (string equals)
where it was supposed to use regex pattern match. Oops. Would have led to
disk vendor id failures in several cases.

Fixes:
1. help, man updates for RAM/Memory data, more clarifications.

2. Refactored RepoData class/package, to make it easier to handle repo string
data, and make it all overall cleaner internally, and enable future extensions
to certain features in inxi that may or may not one day become active.

3. Added to some regex compares \Q$VAR\E to disable regex characters in strings.
I should have used that a long time ago, oh well, better late than never!

4. Found a horrible case were xdpyinfo uses 'preferred' instead of the actual
pixel dimensions, shame on whoever allowed that output!!! shame! Had to add
a workaround to make sure numeric values are present, if not, then use the
fallback, which means, 2x more data parsing to get data that should not
require that, but in this example, it did (an Arch derivative, but it could
be xdpyinfo itself, don't know)>

Enhancements:
1. More fixes on issue #185. Thanks tubecleaner for finding and provding required
data to really solve a set of RAM issues that apply particularly in production
systems. This issue report led to 2 new options: --memory-short, which only
shows a basic RAM report.

Memory:    RAM: total: 31.43 GiB used: 14.98 GiB (47.7%)
           Report: arrays: 1 slots: 4 modules: 2 type: DDR4

And a 2nd, --memory-modules, only shows the occupied slots. This can be
useful in situations where it's a server or vm with a lot of slots, most empty:

Memory:    RAM: total: 31.43 GiB used: 15.44 GiB (49.1%)
           Array-1: capacity: 256 GiB slots: 4 EC: None
           Device-1: DIMM 1 size: 16 GiB speed: 2400 MT/s
           Device-2: DIMM 1 size: 16 GiB speed: 2400 MT/s

Note that both of these options trigger -m, so -m itself is not required.

2. More disk vendors!! The list never ends! Thanks linux-lite hardware database
and users for supplying, and buying/obtaining, apparently every disk known to
mankind.

3. Added fallback XFCE detection, in cases were the system does not have xprop
installed, it's still possible to do a full detection of xfce, including toolkit,
so now inxi does that, one less dependency to detect one more desktop.

4. Added vmwgfx driver to xorg drivers list. Note, I've never actually seen this
in the wild, but I did see it as the kernel reported driver from lspci, so it
may exist.

Unfixed:
1. Issue #187 EnochTheWise (?) did not supply the required debugger data so there
is a RAID ZFS issue that will not get fixed until the required debugger data is
supplied. Please do not waste all our time filing an issue if you have no
intention of actually following through so we can get it fixed.

Note that a key way we get issues here is from Perl errors on the screen, which are
a frequent cause of someone realizing something is wrong. This is why I'm not going
to do a hack fix for the RAID ZFS issue, then the error messages will go away, and
it will likely never get handled. For examples of good, useful, productive issue
reports, and how to do them right: #188 and #185, both of which led to good
improvements in how inxi handles corner cases in those areas.